requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
requests>=2.31.0
python-dateutil>=2.8.0
sqlite3
email-validator>=2.0.0
python-dotenv>=1.0.0
amadeus>=8.0.0
beautifulsoup4>=4.12.0
selenium>=4.15.0